requirements.txt
asgiref==3.8.1
black==26.1.0
click==8.3.1
Django==5.0.6
django-jalali==6.2.1
django-jalali-date==1.1.0
django-unfold==0.55.2
et-xmlfile==1.1.0
gunicorn==22.0.0
isort==7.0.0
jalali_core==1.0.0
jdatetime==5.0.0
mypy_extensions==1.1.0
numpy==1.26.4
openpyxl==3.1.2
packaging==26.0
pandas==2.2.2
pathspec==1.0.4
platformdirs==4.9.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytokens==0.4.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tomli==2.4.0
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.7.0
